Working react-hot-loader. Pending to update readme
