+++ 10 HotLoader/.babelrc
@@ -1,5 +1,13 @@
 {
   "presets": [
-    "env",
+    [
+      "env",
+      {
+        "modules": false
+      }
+    ]
+  ],
+  "plugins": [
+    "react-hot-loader/babel"
   ]
 }
