+++ 10 HotLoader/package.json
@@ -10,13 +10,16 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
+    "@types/core-js": "^0.9.41",
     "@types/react": "^15.0.21",
     "@types/react-dom": "^0.14.23",
+    "@types/react-hot-loader": "^3.0.1",
     "@types/react-redux": "^4.4.38",
-    "@types/react-router": "^2.0.38",
+    "@types/react-router": "^3.0.11",
     "@types/react-router-redux": "^4.0.34",
     "@types/redux-thunk": "^2.1.0",
     "@types/toastr": "^2.1.33",
+    "@types/webpack-env": "^1.13.0",
     "awesome-typescript-loader": "^3.1.2",
     "babel-core": "^6.24.0",
     "babel-preset-env": "^1.3.2",
@@ -24,6 +27,7 @@
     "extract-text-webpack-plugin": "^2.1.0",
     "file-loader": "^0.10.1",
     "html-webpack-plugin": "^2.24.0",
+    "react-hot-loader": "^3.0.0-beta.6",
     "style-loader": "^0.16.0",
     "ts-loader": "^2.0.3",
     "typescript": "^2.0.6",
@@ -32,7 +36,6 @@
     "webpack-dev-server": "^2.4.2"
   },
   "dependencies": {
-    "@types/core-js": "^0.9.41",
     "bootstrap": "^3.3.7",
     "core-js": "^2.4.1",
     "jquery": "^3.2.1",
