+++ 10 HotLoader/src/main.tsx
@@ -1,40 +1,22 @@
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
-import { createStore, applyMiddleware, compose } from 'redux';
-import { Router, Route, IndexRoute, hashHistory } from 'react-router';
-import { syncHistoryWithStore } from 'react-router-redux'
-import reduxThunk from 'redux-thunk';
-import { Provider } from 'react-redux';
-import {reducers} from './reducers'
-import {App} from './app';
-import {LoginContainer} from './pages/login';
-import {StudentListContainer} from './pages/student-list';
-import {StudentDetailContainer} from './pages/student-detail';
+import {AppProvider} from './provider';
+import {AppContainer} from 'react-hot-loader';
 
+const render = (Component) => {
+  ReactDOM.render(
+    <AppContainer>
+      <Component />
+    </AppContainer>,
+    document.getElementById('root')
+  );
+};
 
-let store = createStore(
-  reducers,
-  compose(
-    applyMiddleware(reduxThunk),
-    window['devToolsExtension'] ? window['devToolsExtension']() : f => f
-  )  
-);
+render(AppProvider);
 
-const history = syncHistoryWithStore(hashHistory, store);
-
-ReactDOM.render(
-  <Provider store={store}>
-    <div>
-      <Router history={history}>
-        <Route path="/" component={App}>
-          <IndexRoute component={LoginContainer}/>
-          <Route path="login" component={LoginContainer}/>
-          <Route path="student-list" component={StudentListContainer}/>
-          <Route path="student-detail/:id" component={StudentDetailContainer}/>
-        </Route>
-      </Router>
-    </div>
-  </Provider>,
-  document.getElementById('root')
-);
+if(module.hot) {
+  module.hot.accept('./provider', () => {
+    render(AppProvider);
+  });
+}
