+++ 10 HotLoader/src/pages/login/components/header.tsx
@@ -1,9 +1,9 @@
 import * as React from "react"
 
 export const Header = () => {
-  return (    
-  	   <div className="panel-heading">
-  	     <h3 className="panel-title">Please sign in</h3>
-  	   </div>
-  );
+  return (
+		<div className="panel-heading">
+			<h3 className="panel-title">Please sign in</h3>
+		</div>
+	);
 }
