+++ 10 HotLoader/src/provider.tsx
@@ -1,0 +1,13 @@
+import * as React from 'react';
+import * as ReactDOM from 'react-dom';
+import { Provider } from 'react-redux';
+import {store} from './store';
+import {AppRouter} from './router';
+
+export const AppProvider = () => {
+  return (
+    <Provider store={store}>
+      <AppRouter />
+    </Provider>
+  );
+}
